test: fix minor test bugs

+++ evm/test/CatalystFactory/DeployVault.t.sol
@@ -84,7 +84,7 @@ contract TestDeployVault is TestCommon {
     }
 
     function test_deploy_3_token_amplified(uint16[3] memory weights_, uint64 amplificationPercentage) external {
-        uint64 amplification = 10**18 * amplificationPercentage / type(uint64).max;
+        uint64 amplification = uint64(uint256(10**18) * uint256(amplificationPercentage) / uint256(type(uint64).max));
         vm.assume(amplification < 10**18);
         vm.assume(0 < amplification);
         vm.assume(weights_[0] > 0);
@@ -107,7 +107,7 @@ contract TestDeployVault is TestCommon {
     }
 
     function test_deploy_2_token_amplified(uint16[2] memory weights_, uint64 amplificationPercentage) external {
-        uint64 amplification = 10**18 * amplificationPercentage / type(uint64).max;
+        uint64 amplification = uint64(uint256(10**18) * uint256(amplificationPercentage) / uint256(type(uint64).max));
         vm.assume(amplification < 10**18);
         vm.assume(0 < amplification);
         vm.assume(weights_[0] > 0);
